# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Region {
  id: string;
  name: string;
  type: string;
  processedSqKm: number;
  parcelCount: number;
  updatedAt: string;
}

export interface ProcessingRun {
  id: string;
  name: string;
  regionId: string;
  regionName: string;
  status: string;
  progress: number;
  currentStep: string;
  steps: string[];
  startedAt: string;
}

export interface Activity {
  id: string;
  title: string;
  detail: string;
  time: string;
  tone: string;
}

export interface DashboardSummary {
  areaProcessedSqKm: number;
  parcelsExtracted: number;
  topologyErrors: number;
  encroachments: number;
  accuracyScore: number;
  activeRun: ProcessingRun | null;
  recentActivity: Activity[];
}

export interface Parcel {
  id: string;
  ulpin: string;
  regionId: string;
  regionName: string;
  areaSqM: number;
  ownership: string;
  confidence: number;
  status: string;
  geometry: number[][];
  updatedAt: string;
}

export interface ParcelUpdate {
  status?: string;
  geometry?: number[][];
}

export interface ProcessingRunInput {
  regionId: string;
  dataset: string;
}

export interface TopologyReport {
  scannedAt: string;
  errorCount: number;
  fixedCount: number;
  overlapCount: number;
  sliverCount: number;
  ringCount: number;
  status: string;
}

export interface ChangeDetection {
  id: string;
  title: string;
  regionName: string;
  category: string;
  areaSqM: number;
  detectedAt: string;
  severity: string;
}

export interface ExportInput {
  format: string;
  regionId: string;
}

export interface ExportJob {
  id: string;
  format: string;
  regionId: string;
  status: string;
  createdAt: string;
  fileName: string;
}

export type AnalysisSourceType = typeof AnalysisSourceType[keyof typeof AnalysisSourceType];


export const AnalysisSourceType = {
  image: 'image',
  video: 'video',
} as const;

export type AnalysisStatus = typeof AnalysisStatus[keyof typeof AnalysisStatus];


export const AnalysisStatus = {
  processing: 'processing',
  needs_review: 'needs_review',
  verified: 'verified',
  needs_attention: 'needs_attention',
} as const;

export interface Analysis {
  id: string;
  title: string;
  sourceType: AnalysisSourceType;
  sourceName: string;
  status: AnalysisStatus;
  createdAt: string;
  observationCount: number;
  verifiedObservationCount: number;
  hasGeospatialMetadata: boolean;
  /** @nullable */
  captureDate: string | null;
  modelLabel: string;
}

export interface SentinelOverview {
  totalAnalyses: number;
  pendingReview: number;
  verified: number;
  needsAttention: number;
  latestAnalysis: Analysis | null;
}

export type AnalysisInputSourceType = typeof AnalysisInputSourceType[keyof typeof AnalysisInputSourceType];


export const AnalysisInputSourceType = {
  image: 'image',
  video: 'video',
} as const;

export type AnalysisInputStatus = typeof AnalysisInputStatus[keyof typeof AnalysisInputStatus];


export const AnalysisInputStatus = {
  processing: 'processing',
  needs_review: 'needs_review',
  verified: 'verified',
  needs_attention: 'needs_attention',
} as const;

export type ObservationInputReviewStatus = typeof ObservationInputReviewStatus[keyof typeof ObservationInputReviewStatus];


export const ObservationInputReviewStatus = {
  unreviewed: 'unreviewed',
  confirmed: 'confirmed',
  dismissed: 'dismissed',
} as const;

export interface ObservationInput {
  label: string;
  category: string;
  confidence: number;
  reviewStatus: ObservationInputReviewStatus;
  evidenceNote: string;
  /** @nullable */
  timestampSeconds: number | null;
  sourceMedia: string;
}

export interface AnalysisInput {
  title: string;
  sourceType: AnalysisInputSourceType;
  sourceName: string;
  status: AnalysisInputStatus;
  observationCount: number;
  hasGeospatialMetadata: boolean;
  /** @nullable */
  captureDate: string | null;
  modelLabel: string;
  observations?: ObservationInput[];
}

export type MultimodalAnalysisInputSourceType = typeof MultimodalAnalysisInputSourceType[keyof typeof MultimodalAnalysisInputSourceType];


export const MultimodalAnalysisInputSourceType = {
  image: 'image',
  video: 'video',
} as const;

export interface MediaFrameInput {
  data: string;
  mimeType: string;
  timestampSeconds: number;
}

export interface MultimodalAnalysisInput {
  title: string;
  sourceType: MultimodalAnalysisInputSourceType;
  sourceName: string;
  hasGeospatialMetadata: boolean;
  /** @nullable */
  captureDate: string | null;
  /** Base64-encoded image or representative media payload without the data URI prefix. */
  mediaData: string;
  mediaMimeType: string;
  frames?: MediaFrameInput[];
}

export type ObservationReviewStatus = typeof ObservationReviewStatus[keyof typeof ObservationReviewStatus];


export const ObservationReviewStatus = {
  unreviewed: 'unreviewed',
  confirmed: 'confirmed',
  dismissed: 'dismissed',
} as const;

export interface Observation {
  id: string;
  label: string;
  category: string;
  confidence: number;
  reviewStatus: ObservationReviewStatus;
  evidenceNote: string;
  /** @nullable */
  timestampSeconds: number | null;
  sourceMedia: string;
}

export type AnalysisDetail = Analysis & {
  observations: Observation[];
};

export type VerificationInputStatus = typeof VerificationInputStatus[keyof typeof VerificationInputStatus];


export const VerificationInputStatus = {
  needs_review: 'needs_review',
  verified: 'verified',
  needs_attention: 'needs_attention',
} as const;

export interface VerificationInput {
  status: VerificationInputStatus;
}

export interface AnalysisReport {
  generatedAt: string;
  title: string;
  disclaimer: string;
  analysis: AnalysisDetail;
}

export type ListParcelsParams = {
regionId?: string;
status?: string;
};