# lib/api-zod/src/generated/types/analysisInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AnalysisInputSourceType } from './analysisInputSourceType';
import type { AnalysisInputStatus } from './analysisInputStatus';
import type { ObservationInput } from './observationInput';

export interface AnalysisInput {
  title: string;
  sourceType: AnalysisInputSourceType;
  sourceName: string;
  status: AnalysisInputStatus;
  observationCount: number;
  hasGeospatialMetadata: boolean;
  /** @nullable */
  captureDate: string | null;
  modelLabel: string;
  observations?: ObservationInput[];
}