# lib/api-zod/src/generated/types/analysisReport.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AnalysisDetail } from './analysisDetail';

export interface AnalysisReport {
  generatedAt: string;
  title: string;
  disclaimer: string;
  analysis: AnalysisDetail;
}