# lib/api-zod/src/generated/types/observation.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ObservationReviewStatus } from './observationReviewStatus';

export interface Observation {
  id: string;
  label: string;
  category: string;
  confidence: number;
  reviewStatus: ObservationReviewStatus;
  evidenceNote: string;
  /** @nullable */
  timestampSeconds: number | null;
  sourceMedia: string;
}