# lib/api-zod/src/generated/types/verificationInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { VerificationInputStatus } from './verificationInputStatus';

export interface VerificationInput {
  status: VerificationInputStatus;
}